Update GameConfig.cfg
Changed the name field on line 3 from "Nazi Zombies: Portable" to "Nazi Zombies Portable"

When the name field includes a colon (:), it will mess up TrenchBroom when it tries to create a directory with this name (on Windows 11). It attempts to create a directory with a colon in it which Windows doesnt allow - this makes it not create a directory. This causes you to be unable to save an engine after clicking "Configure Engines". By changing the name and removing the colon (:), it fixes this issue.